pyproject.toml
[tool.poetry]
name = "currency-rates"
version = "0.1.0"
description = ""
authors = ["Selivonchyk <selivonchik_n@rocketdata.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
djangorestframework = "^3.15.1"
drf-spectacular = "^0.27.2"
requests = "^2.32.3"
isort = "^5.13.2"
flake8 = "^7.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
